melancholic_poet_qlora/requirements.txt
torch>=2.6.0
transformers>=4.53.0,<5.0.0
accelerate>=1.1.0
peft>=0.18.0
datasets>=3.0.0
bitsandbytes>=0.45.0; platform_system == "Linux"
pyyaml>=6.0.1
safetensors>=0.4.5
sentencepiece>=0.2.0
numpy>=1.26.0
tqdm>=4.66.0
